Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 int main() {
-double r , area;
+float r , area;
 scanf("%lf" ,&r);
 area =M_PI *r *r;
 printf("Area: %.2lf", area);
